Move extensions to VC files and add tests
